Introduce semicolon to the gang of forms.
It is now possible to search for cities such as Biel;Bienne. Also the search itself only relies on the entered zipcode instead of the cityname. To be a bit more precise, the regex pattern in SearchForm.java allows the semicolon as part of the city string. So by choosing citynames that contain a semicolon SearchForm does not complain anymore. Also, it is now possible to alter the cityname in the search form input. For example, one could choose '8000 - Zürich' and mess with the name: '8000 - Zürrriiii' or '8000 - Schnöseldörfli'. This is possible because the collection of matching cities in AdService#queryResults is now solely based on the zipcode instead of the cityname. Nothing else has changed.
Thinking about placeAd, it would be not a stupid idea to restrict publishers to citynames that actually exist.