Nicify select inputs in signup.jsp
Select inputs do not have a title anymore. The title was confusing, not only for the pointer cursor declared to it. Instead the select shows a placeholder. The placeholder is just a disabled input option, meaning that, leaving the select input unchanged when submitting, an error will be triggered. The controller properties are therefore set to @NotNull with a custom error message.