Wurschtl oauth setup (in fact just recv. google user details and doing random shti with it)
It is working as is. But, all this should be handled completely diffr. There needs to be config in springSecurity.xml and one should work with the token provided by google. There is a lot about it in the internetz. Right now i do not have time dive deeper into it. Maybe on Tuesday. The problem with saving user details: simple typos, wrong typing and a user role for the google user that is secured by spring sec, which is conf'd to allow access to users with this role only with a password... which a google user does not have. I set a random user role for the google user just for demo.