Make alerts page work. Refurb as well.
When submitting a faulty form, the existing alerts list was not shown. Bad. I had to change the controller a bit; instead of returning a new alerts model returnprepareAlertPage(principal, false, alertForm) (NB: false set, so the existing alerts stay visible). Also, 'St. Gallen' could not be submitted due to the dot in it. I corrected that in all forms where autocompletion for city field is used.